// src/pages/hr/Employee Master/employee-hr.types.ts
export type TEmployeeHr = TPersnolHr & TPayrollHr & TPassportHr & TContractHr & TSponsorHr & TIsuranceHr & TILPHr & TAirfareHr;

export type TTableEmployeeHr = TPersnolHr &
  TPayrollHr &
  TPassportHr &
  TContractHr &
  TSponsorHr &
  TIsuranceHr &
  TILPHr &
  TAirfareHr &
  TTableHr;
//Persnole information
export type TPersnolHr = {
  company_code:string;
  employer_code:string;
  section_code: string;
  dept_code: string;
  div_code: string;
  emp_photo: string;
  employee_id: string,
  employee_code: string;
  alternate_id: string;
  rpt_name: string;
  grade_code: string;
  desg_code: string;
  labour_desg_code: string;
  category_code: string;
  birth_date: Date;
  join_date: Date;
  probation_end_date: Date;
  probation_confirm_date: Date;
  emp_status: string;
  country_code: string;
  actions?: any;
};

export type TPayrollHr = {
  include_in_payroll: string;
  payroll_start_date: Date;
  payment_mode: string;
  company_bank_code: string;
  salary_acct_no: string;
  salary_bank_code: string;
  currency_id: string;
  exch_rate: number;
  emp_iban_no: string;
  actions?: any;
};

export type TPassportHr = {
  ppt_no: string;
  ppt_name: string;
  ppt_country: string;
  ppt_status: string;
  ppt_valid_from: Date;
  ppt_valid_to: Date;
  passport_with: string;
};

export type TContractHr = {
  contract_type: string;
  contract_start_date: Date;
  contract_end_date: Date | null;
  contract_renewable: string;
  contract_type_desc?: string;
};

export type TSponsorHr = {
  sponsor_id: string | number;
  visa_type: string;
  visa_valid_from: Date;
  visa_valid_to: Date | null;
};

export type TIsuranceHr = {
  ins_card_no: string;
  ins_card_issue_dt: Date;
  ins_card_exp_dt: Date;
  ins_card_type: string;
};

export type TILPHr = {
  labourcard_no: string;
  pasi_no: string;
  labourcard_valid_from: Date;
  labourcard_valid_to: Date;
  labourcard_status: string;
};

export type TAirfareHr = {
  airport_code: string;
  ticket_eligibility: string;
  ticket_dpend_adult: number;
  ta_no: number;
  tc_no: number;
  ti_no: number;
  ticket_eligible_period: number;
};